renamed search field to search input
