+++ src/components/form-field/form-field.tsx
@@ -11,7 +11,7 @@ import { Listbox } from "./listbox/listbox";
 import { MultiCombobox } from "./multi-combobox/multi-combobox";
 import { SingleCombobox } from "./single-combobox/single-combobox";
 import { FormFieldGroup } from "./form-field-group";
-import { SearchInput } from "./search-field/search-field";
+import { SearchInput } from "./search-input/search-input";
 
 interface FormFieldProps {
     children: React.ReactNode;
